Rename shell scripts
The usual convention is have no file name extension for executable
scripts, and to name library scripts according to their shell (here:
.bash). The rationale for the first part is that users do not need to
konw whether a program is a shell script or not, and this allows one to
rewrite the program in a different language (interpreted or compiled)
without having to update anything that refers to the program name
(documentation, crontab, etc.).

In this case, the file name extensions were also misleading, since the
scripts have a Bash shebang and use Bash features. If a user, based on
the file name extension, had tried to run a script as `sh all.sh`, it
would not have worked.

Citation: Google Shell Style Guide,
https://google.github.io/styleguide/shell.xml#File_Extensions